[CHORE] Combine buildings and basket/chest buttons
