+++ src/features/island/buildings/Tutorial.tsx
@@ -1,33 +1,21 @@
 import React, { useState } from "react";
 import { Button } from "components/ui/Button";
-import { CloseButtonPanel } from "features/game/components/CloseablePanel";
-
-import { Equipped } from "features/game/types/bumpkin";
 import { SUNNYSIDE } from "assets/sunnyside";
+import { PIXEL_SCALE } from "features/game/lib/constants";
 
 interface Props {
   onClose: () => void;
 }
 
-type Pages = 1 | 2;
+type Pages = 0 | 1;
 
 export const Tutorial: React.FC<Props> = ({ onClose }) => {
-  const [pageNumber, setPageNumber] = useState<Pages>(1);
-
-  const bumpkinParts: Partial<Equipped> = {
-    body: "Beige Farmer Potion",
-    hair: "Blacksmith Hair",
-    pants: "Brown Suspenders",
-    shirt: "Yellow Farmer Shirt",
-    tool: "Hammer",
-    background: "Farm Background",
-    shoes: "Black Farmer Boots",
-  };
+  const [pageNumber, setPageNumber] = useState<Pages>(0);
 
   const PageOne = () => {
     return (
       <>
-        <div className="space-y-3 text-sm px-1 mb-3">
+        <div className="space-y-3 text-sm px-2 mb-3">
           <p>
             This menu will show you the levels required to unlock new buildings.
           </p>
@@ -36,7 +24,7 @@ export const Tutorial: React.FC<Props> = ({ onClose }) => {
             level.
           </p>
         </div>
-        <Button onClick={() => setPageNumber(2)}>Next</Button>
+        <Button onClick={() => setPageNumber(1)}>Next</Button>
       </>
     );
   };
@@ -44,7 +32,7 @@ export const Tutorial: React.FC<Props> = ({ onClose }) => {
   const PageTwo = () => {
     return (
       <>
-        <div className="space-y-3 text-sm px-1 mb-3">
+        <div className="space-y-3 text-sm px-2 mb-3">
           <p>
             Buildings are an important way to progress through the game as they
             will help you to expand and evolve.
@@ -55,7 +43,7 @@ export const Tutorial: React.FC<Props> = ({ onClose }) => {
           </p>
         </div>
         <div className="flex space-x-1">
-          <Button onClick={() => setPageNumber(1)}>Back</Button>
+          <Button onClick={() => setPageNumber(0)}>Back</Button>
           <Button onClick={onClose}>Got it</Button>
         </div>
       </>
@@ -63,16 +51,18 @@ export const Tutorial: React.FC<Props> = ({ onClose }) => {
   };
 
   return (
-    <CloseButtonPanel
-      onClose={onClose}
-      title="Welcome to Buildings"
-      bumpkinParts={bumpkinParts}
-    >
-      <div className="w-full mb-3 flex justify-center">
-        <img src={SUNNYSIDE.icons.hammer} className="h-6 md:h-7" />
+    <>
+      <div className="w-full flex justify-center">
+        <img
+          src={SUNNYSIDE.icons.hammer}
+          className="my-2"
+          style={{
+            width: `${PIXEL_SCALE * 13}px`,
+          }}
+        />
       </div>
-      {pageNumber === 1 && PageOne()}
-      {pageNumber === 2 && PageTwo()}
-    </CloseButtonPanel>
+      {pageNumber === 0 && PageOne()}
+      {pageNumber === 1 && PageTwo()}
+    </>
   );
 };
