+++ src/features/island/hud/Hud.tsx
@@ -3,7 +3,6 @@ import { Balance } from "components/Balance";
 import { useActor } from "@xstate/react";
 import { Context } from "features/game/GameProvider";
 import { Settings } from "./components/Settings";
-import { Buildings } from "../buildings/Buildings";
 import { Inventory } from "./components/inventory/Inventory";
 import { PlaceableController } from "features/farming/hud/components/PlaceableController";
 import { BumpkinProfile } from "./components/BumpkinProfile";
@@ -28,18 +27,6 @@ export const Hud: React.FC<{ isFarming: boolean }> = ({ isFarming }) => {
       aria-label="Hud"
       className="absolute z-40"
     >
-      {isEditing ? (
-        <PlaceableController />
-      ) : (
-        <>
-          <Balance balance={gameState.context.state.balance} />
-          {landId && <LandId landId={landId} />}
-          <Buildings />
-          <Save />
-          <BumpkinProfile />
-          <Settings isFarming={isFarming} />
-        </>
-      )}
       <div hidden={isEditing}>
         <Inventory
           state={gameState.context.state}
@@ -55,6 +42,17 @@ export const Hud: React.FC<{ isFarming: boolean }> = ({ isFarming }) => {
           isFarming={isFarming}
         />
       </div>
+      {isEditing ? (
+        <PlaceableController />
+      ) : (
+        <>
+          <Balance balance={gameState.context.state.balance} />
+          {landId && <LandId landId={landId} />}
+          <Save />
+          <BumpkinProfile />
+          <Settings isFarming={isFarming} />
+        </>
+      )}
     </div>
   );
 };
