+++ src/features/island/hud/components/inventory/Inventory.tsx
@@ -47,38 +47,62 @@ export const Inventory: React.FC<Props> = ({
   };
 
   return (
-    <div
-      className="flex flex-col items-center fixed z-50"
-      style={{
-        right: `${PIXEL_SCALE * 3}px`,
-        top: isFarming ? `${PIXEL_SCALE * 50}px` : `${PIXEL_SCALE * 25}px`,
-      }}
-    >
+    <>
       <div
-        onClick={() => setIsOpen(true)}
-        className="relative flex z-50 cursor-pointer hover:img-highlight"
+        className="flex flex-col items-center fixed z-50"
         style={{
-          marginLeft: `${PIXEL_SCALE * 2}px`,
-          marginBottom: `${PIXEL_SCALE * 25}px`,
-          width: `${PIXEL_SCALE * 22}px`,
+          right: `${PIXEL_SCALE * 3}px`,
+          top: `${PIXEL_SCALE * 25}px`,
         }}
       >
-        <img
-          src={SUNNYSIDE.ui.round_button}
-          className="absolute"
+        <div
+          onClick={() => setIsOpen(true)}
+          className="relative flex z-50 cursor-pointer hover:img-highlight"
           style={{
+            marginLeft: `${PIXEL_SCALE * 2}px`,
+            marginBottom: `${PIXEL_SCALE * 25}px`,
             width: `${PIXEL_SCALE * 22}px`,
           }}
-        />
-        <img
-          src={SUNNYSIDE.icons.basket}
-          className="absolute"
-          style={{
-            top: `${PIXEL_SCALE * 5}px`,
-            left: `${PIXEL_SCALE * 5}px`,
-            width: `${PIXEL_SCALE * 12}px`,
-          }}
-        />
+        >
+          <img
+            src={SUNNYSIDE.ui.round_button}
+            className="absolute"
+            style={{
+              width: `${PIXEL_SCALE * 22}px`,
+            }}
+          />
+          <img
+            src={SUNNYSIDE.icons.basket}
+            className="absolute"
+            style={{
+              top: `${PIXEL_SCALE * 5}px`,
+              left: `${PIXEL_SCALE * 5}px`,
+              width: `${PIXEL_SCALE * 12}px`,
+            }}
+          />
+        </div>
+
+        {!limitedInventory && (
+          <div
+            className="flex flex-col items-center"
+            style={{
+              marginRight: `${PIXEL_SCALE * -3}px`,
+            }}
+          >
+            {shortcuts.map((item, index) => (
+              <Box
+                key={index}
+                isSelected={index === 0}
+                image={ITEM_DETAILS[item]?.image}
+                secondaryImage={ITEM_DETAILS[item]?.secondaryImage}
+                count={state.inventory[item]?.sub(
+                  state.collectibles[item as CollectibleName]?.length ?? 0
+                )}
+                onClick={() => handleBasketItemClick(item)}
+              />
+            ))}
+          </div>
+        )}
       </div>
 
       <InventoryItemsModal
@@ -93,28 +117,6 @@ export const Inventory: React.FC<Props> = ({
         isSaving={isSaving}
         isFarming={isFarming}
       />
-
-      {!limitedInventory && (
-        <div
-          className="flex flex-col items-center"
-          style={{
-            marginRight: `${PIXEL_SCALE * -3}px`,
-          }}
-        >
-          {shortcuts.map((item, index) => (
-            <Box
-              key={index}
-              isSelected={index === 0}
-              image={ITEM_DETAILS[item]?.image}
-              secondaryImage={ITEM_DETAILS[item]?.secondaryImage}
-              count={state.inventory[item]?.sub(
-                state.collectibles[item as CollectibleName]?.length ?? 0
-              )}
-              onClick={() => handleBasketItemClick(item)}
-            />
-          ))}
-        </div>
-      )}
-    </div>
+    </>
   );
 };
